refactor: organize servers into `listeners` and `senders` folders
