+++ portal/ui/globals.py
@@ -1,4 +1,4 @@
-from ..server.managers.connection_manager import ConnectionManager
+from ..server.connection_manager import ConnectionManager
 
 MODAL_OPERATORS = {}
 CONNECTION_MANAGER = ConnectionManager()
